study-pipeline/scripts/dashboard/requirements.txt
streamlit>=1.30
plotly>=5.0
pandas>=2.0
ruamel.yaml>=0.18
